cloud_functions/data_analysis/requirements.txt
functions-framework==3.*
google-cloud-bigquery == 3.12.0
